requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-multipart==0.0.20
pydantic==2.11.0
pydantic-settings==2.7.1
python-magic==0.4.27
pefile==2023.2.7
pyelftools==0.31
flare-capa==9.3.1
celery==5.4.0
redis==5.2.0
httpx==0.27.0
